Refactor code to avoid build tags to pick network.
Now, the network is an interface and it can be picked with a flag,
rather than build tags.

This is simpler and more "standard". It also enables full testing to be
done without hardcoding all the build tag combinations.
