+++ .travis.yml
@@ -14,10 +14,6 @@ install:
   - # Do nothing. This is needed to prevent default install action "go get -t -v ./..." from happening here (we want it to happen inside script step).
 script:
   - go get -t -v ./...
-  - go get -t -v -tags=tcp ./...
-  - go get -t -v -tags=chan ./...
   - diff -u <(echo -n) <(gofmt -d -s .)
   - go tool vet .
   - go test -v ./... # -race
-  - go test -v -tags=tcp ./... # -race
-  - go test -v -tags=chan ./... # -race
